Added deleteBook(id) in BookItem
I have added a deleteBook method that takes the id of the book as a parameter and emits the delete-book event with this id value. We have also added a click event listener to the delete button that calls the deleteBook method with the id of the book as its parameter when clicked.