Added some features in SearchBar
The new features are:

showForm(): Emits an event show-form when the "Add book" button is clicked.

clearSearch(): Empties the input field and emits an event search with an empty string when the "Clear" button is clicked.

searchBook(): Emits an event search with the content of the search field every time it is modified. The v-model directive is used to bind the input element to the searchTerm data property.