Added sortBy, orderBy and favourite in FIlterBar
The sortBy() method handles emitting an event called sort-items with the value of the selected option in the Sort By select element. 

Similarly, the orderBy() method emits an event called order-items with the value of the selected option in the Order By select element. 

Finally, the favourite() method handles emitting an event called favourite-items with the value of the Favourite checkbox. The template has also been updated to call these methods on the appropriate events.